Изменяет компонент CardsList  для использования на всех страницах
