+++ src/pages/main-page/main-page.tsx
@@ -1,6 +1,6 @@
 import { Helmet } from 'react-helmet-async';
 
-import CardsList from '../../components/cards-list/cards-list';
+import CardsSection from '../../components/cards-section/cards-section';
 import CitiesList from '../../components/cities-list/cities-list';
 import Header from '../../components/header/header';
 import Map from '../../components/map/map';
@@ -29,7 +29,7 @@ export default function MainPage({ placesCount, offers }: MainPageProps): JSX.El
         </div>
         <div className="cities">
           <div className="cities__places-container container">
-            <CardsList placesCount={placesCount} offers={offers} />
+            <CardsSection placesCount={placesCount} offers={offers} />
             <div className="cities__right-section">
               <Map
                 city={offers[0].city}
